@@ -3,8 +3,8 @@ using SocialMedia.Domain.Exceptions;
 using SocialMedia.Domain.Models;
 using SocialMedia.Persistence.SqlServer.Extensions;
 using SocialMedia.Persistence.SqlServer.Models;
+using SocialMedia.TestUtilities;
 using SocialMedia.TestUtilities.Builders;
-using static SocialMedia.TestUtilities.Builders.PostConfiguration;
 
 namespace SocialMedia.Persistence.SqlServer.Tests.Extensions
 {
@@ -12,11 +12,11 @@ namespace SocialMedia.Persistence.SqlServer.Tests.Extensions
     {
         private readonly PostBuilder postBuilder = new();
 
-        [Fact]
+        [Fact, UseMappingContextScope]
         public void ToPostInfo_ReturnsExpectedPostInfo()
         {
             var post = postBuilder.CreatePost();
-            var postData = post.ToPostData(MappingBehavior.IncludeUser);
+            var postData = post.ToPostData();
 
             var result = postData.ToPostInfo();
 
@@ -44,22 +44,6 @@ namespace SocialMedia.Persistence.SqlServer.Tests.Extensions
 
             action.Should().Throw<ModelMappingException<PostData, PostInfo>>()
                 .WithMessage($"*Missing {nameof(PostData.User)} value*");
-        }
-    }
-
-    public class UserExtensionsTests
-    {
-        private readonly UserBuilder userBuilder = new();
-
-        [Fact]
-        public void ToUser_ReturnsExpectedUserInfo()
-        {
-            var user = userBuilder.CreateUser();
-            var userData = user.ToUserData();
-
-            var result = userData.ToUser();
-
-            result.Should().Be(user.ToUser());
         }
     }
 }
