@@ -1,6 +1,5 @@
 using SocialMedia.Domain.Models;
-using SocialMedia.Persistence.SqlServer.Models;
-using SocialMedia.TestUtilities.Exceptions;
+using SocialMedia.TestUtilities.Models;
 
 namespace SocialMedia.TestUtilities.Builders
 {
@@ -29,84 +28,6 @@ namespace SocialMedia.TestUtilities.Builders
         {
             public required string UserId { get; init; }
             public required string Name { get; init; }
-        }
-    }
-
-    public record PostConfiguration
-    {
-        public PostId? Id { get; private set; }
-        public UserConfiguration? Author { get; private set; }
-        public string? Text { get; private set; }
-        public DateTime? Created { get; private set; }
-
-        public PostConfiguration WithId(PostId? id)
-        {
-            Id = id;
-            return this;
-        }
-
-        public PostConfiguration WithAuthor(UserConfiguration? author)
-        {
-            Author = author;
-            return this;
-        }
-
-        public PostConfiguration WithText(string? text)
-        {
-            Text = text;
-            return this;
-        }
-
-        public PostConfiguration WithCreated(DateTime? created)
-        {
-            Created = created;
-            return this;
-        }
-
-        public Post ToPost()
-        {
-            return new Post
-            {
-                Id = Id ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Id)),
-                AuthorUserId = Author?.Id ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Author)),
-                Text = Text ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Text)),
-                Created = Created ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Created))
-            };
-        }
-
-        public PostInfo ToPostInfo()
-        {
-            return new PostInfo
-            {
-                Id = Id ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Id)),
-                Author = Author?.Name ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Author)),
-                Text = Text ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Text)),
-                Created = Created ?? throw new NullMappingException<PostConfiguration, Post>(nameof(Created))
-            };
-        }
-
-        public PostData ToPostData(MappingBehavior mappingBehavior = MappingBehavior.Default)
-        {
-            return new PostData
-            {
-                Id = Id?.Value ?? throw new NullMappingException<PostConfiguration, PostData>(nameof(Id)),
-                AuthorUserId = Author?.Id?.Value ?? throw new NullMappingException<PostConfiguration, PostData>(nameof(Author.Id)),
-                Created = Created ?? throw new NullMappingException<PostConfiguration, PostData>(nameof(Created)),
-                Content = new PostContentData
-                {
-                    PostId = Id?.Value ?? throw new NullMappingException<PostConfiguration, PostData>(nameof(PostData.Id)),
-                    Text = Text ?? throw new NullMappingException<PostConfiguration, PostData>(nameof(PostData.Id)),
-                },
-                User = mappingBehavior.HasFlag(MappingBehavior.IncludeUser) ?
-                    Author.ToUserData() : null
-            };
-        }
-
-        [Flags]
-        public enum MappingBehavior
-        {
-            Default = 0,
-            IncludeUser = 1
         }
     }
 }
